add sanity check subcommand to server subcommand
