+++ lib/crowbar/client/command/server/check.rb
@@ -19,36 +19,38 @@ module Crowbar
     module Command
       module Server
         #
-        # Implementation for the server status command
+        # Implementation for the server check command
         #
-        class Status < Base
+        class Check < Base
           include Mixin::Format
           include Mixin::Filter
 
           def request
-            @request ||= Request::Server::Status.new(
+            @request ||= Request::Server::Check.new(
               args
             )
           end
 
           def execute
             request.process do |request|
+              formatter = Formatter::Array.new(
+                format: provide_format,
+                headings: ["Errors"],
+                values: Filter::Array.new(
+                  filter: provide_filter,
+                  values: content_from(request)
+                ).result
+              )
+
               case request.code
               when 200
-                formatter = Formatter::Array.new(
-                  format: provide_format,
-                  headings: ["Errors"],
-                  values: Filter::Array.new(
-                    filter: provide_filter,
-                    values: content_from(request)
-                  ).result
-                )
-
                 if formatter.empty?
                   say "Crowbar is in a sane state"
                 else
                   say formatter.result
                 end
+              when 409
+                err "Could not perform sanity checks: #{request.parsed_response["error"]}"
               else
                 err request.parsed_response["error"]
               end
@@ -58,7 +60,6 @@ module Crowbar
           protected
 
           def content_from(request)
-            return [] if request.body == ""
             [].tap do |row|
               request.parsed_response.each do |child|
                 row.push(
